refactor(condo): DOMA-7635 reused bad-words-next for bad words exclusions
